third_party/WebKit/LayoutTests/xr/xrView_match.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="../xr/resources/xr-device-mocking.js"></script>
<script src="../xr/resources/xr-test-utils.js"></script>
<script src="../xr/resources/test-constants.js"></script>
<canvas id="webgl-canvas"></canvas>

<script>
let fakeDevices = fakeXRDevices();

xr_session_promise_test( (session) => {
  // Session must have a baseLayer or frame requests will be ignored.
  let webglLayer = new XRWebGLLayer(session, gl);
  session.baseLayer = webglLayer;

  return session.requestFrameOfReference("eyeLevel").then((frameOfRef) => new Promise((resolve) =>{
    function onFrame(time, xrFrame) {
      // Ensure that two views are provided.
      assert_not_equals(xrFrame.views, null);
      assert_equals(xrFrame.views.length, 2);

      let leftView = xrFrame.views[0];
      let rightView = xrFrame.views[1];

      // Ensure that the views are the right type.
      assert_true(leftView instanceof XRView);
      assert_true(rightView instanceof XRView);

      // Ensure that they have the expected eye enums.
      assert_equals(leftView.eye, "left");
      assert_equals(rightView.eye, "right");

      // Ensure they have the expected projection matrices.
      assert_not_equals(leftView.projectionMatrix, null);
      assert_not_equals(rightView.projectionMatrix, null);

      let displayLeftEye = fakeDevices["FakeGooglePixelPhone"].leftEye;
      let displayRightEye = fakeDevices["FakeGooglePixelPhone"].rightEye;

      let expectedLeftProjection = perspectiveFromFieldOfView(
          displayLeftEye.fieldOfView, session.depthNear, session.depthFar);
      let expectedRightProjection = perspectiveFromFieldOfView(
          displayRightEye.fieldOfView, session.depthNear, session.depthFar);

      assert_matrices_approx_equal(leftView.projectionMatrix, expectedLeftProjection);
      assert_matrices_approx_equal(rightView.projectionMatrix, expectedRightProjection);

      // Finished test.
      resolve();
    }

    session.requestAnimationFrame(onFrame);
  }));
}, fakeDevices["FakeGooglePixelPhone"], [{ exclusive: true }],
"XRPresentationFrame contains the expected views");

</script>
